feat(studyRegistrationAccessRequest): Reorganized code so DiscoveryForm logic is in a DiscoveryForm folder instead of in pages
