+++ packages/frontend/src/features/DiscoveryForms/StudyRegistrationAccessRequest/FormContentViews.tsx
@@ -26,8 +26,6 @@ export const FormContentViews = ({
   onSubmit,
   isLoading,
 }: FormContentViewsProps) => {
-  console.log('formBody', formBody);
-  console.log('config', config);
   if (isLoading) {
     return (
       <Center h="50vh">
